\name{passive_smoking}
\alias{passive_smoking}
\docType{data}
\title{Dataset on the effect of environmental tobacco smoke}
\description{Effect of environmental tobacco smoke on lung-cancer in lifetime non-smokers.}
\usage{data(passive_smoking)}
\format{
  A data frame with 37 observations on the following 2 variables.
  \describe{
    \item{\code{lnRR}}{Log-relative risk.}
    \item{\code{selnRR}}{Standard error of log-relative risk.}
  }
}
\details{The sample consists of lung cancer patients and controls that were lifelong non-smokers.
The effect of interest is measured by the relative risk of lung cancer according to whether the spouse currently smoked or had never smoked.}
\references{
Hackshaw, A. K., Law, M. R., and Wald, N.J.
The accumulated evidence on lung cancer and environmental tobacco smoke.
\emph{BMJ}, \bold{315}, 980--988.
}
\seealso{This dataset is analyzed in the help file for \code{\link{DearBegg}}.}
\keyword{datasets}

